Implement manual image warping, improve performance
skimage.transform.warp generates the mapping between source and
destination images, then performs the transformation. The first
calculation is the limiting step and takes an order of magnitude
longer than the second. This function generates only the mapping,
relying on the opencv, and allows caching the result (which needs to
be done only once per run in practice).