Use robust keypoint sorting and correct x-y vs row-col
Previously keypoint sorting was done with a crude heuristic that
required the shape of the calibration patter a priori. Now sorting is
done using a kd tree and nearest-neighbor search to sort the
points (vertically) accurately and consistently which should not
result in occasional nonsensical fitting.

There was previously a mix-up between x-y order and row-column order
in the sorting which became more apparent when doing Gaussian
fits. There was a fudge in place -- in the form of redundant lines --
that got it to work well enough. The correct coordinates are being
passed to the fitting function and the redundant lines have been
removed.